Show details, refine Alerts and Selects, show code
This generalizes the String generation for selects and the tree,
makes some of the utils code slightly better and adjusts it to
also show details if so wanted. Folder got refactored, Alerts now
have other texts, index can generate source code.
This also refactors the element's pages to work with the changes
and removes some commented code.